feat: 메인페이지 화면수정 - 배너 크기 변경 및 버튼 추가 - 검색박스 사이즈에 따른 수정

+++ src/main/java/environment/project/controller/AdminController.java
@@ -26,7 +26,7 @@ public class AdminController {
     }
 
     @PostMapping("/admin")
-    public String updateClubApply(@RequestParam Integer clubApplyNo, @RequestParam String method, @RequestParam String refuseReason,Model model){
+    public String updateClubApply(@RequestParam Long clubApplyNo, @RequestParam String method, @RequestParam String refuseReason,Model model){
         ClubApplyDTO dto = new ClubApplyDTO();
         List<ClubApplyDTO> clubApplyList = clubApplyService.selectClubApply();
         model.addAttribute("list",clubApplyList);
